% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet.R
\name{binder.contribution}
\alias{binder.contribution}
\title{Binder contribution between two partitions}
\usage{
binder.contribution(Z1, Z2, a = 1)
}
\arguments{
\item{Z1}{Integer or factor vector: length-n cluster labels for partition 1.}

\item{Z2}{Integer or factor vector: length-n cluster labels for partition 2.}

\item{a}{Parameter for generalized Binder, takes value between 0 and 2, a = 1 by default (regular loss).}
}
\value{
Numeric vector of length n: VI contribution for each observation.
}
\description{
Calculates the individual contributions of observations to the Binder distance between partitions Z1 and Z2.
}
