% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Binder_compute_Rcpp}
\alias{Binder_compute_Rcpp}
\title{Compute "n-invariant version” of Binder distance between two partitions}
\usage{
Binder_compute_Rcpp(c1, c2, K1, K2, a = 1)
}
\arguments{
\item{c1}{An integer vector of cluster labels for $n$ items. Should be 0-indexed.}

\item{c2}{An integer vector of cluster labels for $n$ items. Should be 0-indexed.}

\item{K1}{An integer, specifying the number of unique clusters in \code{cl1}.}

\item{K2}{An integer, specifying the number of unique clusters in \code{cl2}.}

\item{a}{A parameter used in generalized Binder, takes value between 0 and 2 and 1 by default.}
}
\value{
The VI distance between \code{cl1} and \code{cl2}.
}
\description{
Compute "n-invariant version” of Binder distance between two partitions
}
