% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Binder_Rcpp}
\alias{Binder_Rcpp}
\title{Binder distance between two groups of partitions}
\usage{
Binder_Rcpp(cls1, cls2, K1s, K2s, a = 1)
}
\arguments{
\item{cls1}{A matrix of $n$ columns, where each row contains the 0-indexed cluster labels for n items.}

\item{cls2}{A matrix of $n$ columns, where each row contains the 0-indexed cluster labels for n items.}

\item{K1s}{An integer vector, specifying the number of clusters in each row of \code{cls1}. Must have length equal to the number of rows in \code{cls1}.}

\item{K2s}{An integer vector, specifying the number of clusters in each row of \code{cls2}. Must have length equal to the number of rows in \code{cls2}.}
}
\value{
A matrix containing the VI distance between each row of \code{cls1} and \code{cls2}.
}
\description{
Binder distance between two groups of partitions
}
