% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet.R
\name{choose2}
\alias{choose2}
\title{n choose 2
Computes the binomial coefficient "n choose 2", which is the number of ways to choose 2 items from n without regard to the order.}
\usage{
choose2(n)
}
\arguments{
\item{n}{Integer: the number of items to choose from.}
}
\value{
Integer: the number of combinations of 2 items from n.
}
\description{
n choose 2
Computes the binomial coefficient "n choose 2", which is the number of ways to choose 2 items from n without regard to the order.
}
\examples{
choose2(5) # Returns 10, since there are 10 ways to choose 2 items from 5.
}
