% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet.R
\name{cls.meet}
\alias{cls.meet}
\title{Compute the meet}
\usage{
cls.meet(cls.part)
}
\arguments{
\item{cls.part}{Integer matrix of dimension K (number of particles) by n (number of observations),
with each row representing a clustering/partition.}
}
\value{
A list with components:
  \describe{
    \item{cls.m}{Length-n integer vector: cluster assignment for the meet.}
    \item{k.m}{Integer: number of clusters in the meet.}
    \item{cls.sizes.m}{Integer vector: sizes of clusters in the meet.}
  }
}
\description{
Given a set of clusterings/partitions, finds the "meet" across them
(the common refinement, or the finest common coarsening).
}
\examples{
cls.part <- matrix(c(1,1,2,2, 1,2,1,2), nrow = 2, byrow = TRUE)
meet <- WASABI::cls.meet(cls.part)
}
