% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{comb2}
\alias{comb2}
\title{This function computes the number of ways to choose 2 items from x items.}
\usage{
comb2(x)
}
\arguments{
\item{x}{An integer, the number of items.}
}
\value{
The number of ways to choose 2 items from x items, or 0 if x is less than or equal to 1.
}
\description{
This function computes the number of ways to choose 2 items from x items.
}
