% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet.R
\name{eb.wd.contribution}
\alias{eb.wd.contribution}
\title{CExpected Binder Contribution (WASABI approximation)}
\usage{
eb.wd.contribution(output_wvi, Zhat, a = 1)
}
\arguments{
\item{output_wvi}{List: must contain \code{particles} (matrix) and \code{part.weights} (numeric vector).}

\item{Zhat}{Integer vector: partition to compare.}
}
\value{
Numeric vector of length n: weighted expected VI contribution for \code{Zhat} (each element).
}
\description{
Calculates the EB contribution (with expectation taken with respect to the WASABI distribution) of each observation.
}
