% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ggrange_hist}
\alias{ggrange_hist}
\title{Plot mean/range plot and data histogram side-by-side}
\usage{
ggrange_hist(output_wvi, Y)
}
\arguments{
\item{output_wvi}{A list containing at least elements \code{particles} (matrix) from WASABI output.}

\item{Y}{Numeric vector of data.}
}
\value{
If gridExtra is available, NULL (the plots are drawn). Otherwise, a list of plot objects.
}
\description{
Shows ggrange plot and a histogram of the data together.
}
\examples{
\dontrun{
ggrange_hist(output_wvi, Y)
}
}
