% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ggscatter}
\alias{ggscatter}
\title{Scatterplot: clusters vs 1d data, with jitter}
\usage{
ggscatter(cls, Y)
}
\arguments{
\item{cls}{Numeric or factor vector of cluster assignments for each point.}

\item{Y}{Numeric vector of data.}
}
\value{
A ggplot2 plot object.
}
\description{
Makes a scatter plot of data vs cluster assignments, adding random vertical jitter for visual clarity.
}
\examples{
\dontrun{
ggscatter(output_wvi$particles[1,], Y)
}
}
