% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ggscatter_grid2d}
\alias{ggscatter_grid2d}
\title{Grid of 2D cluster scatterplots from multiple particles}
\usage{
ggscatter_grid2d(output_wvi, Y, title = NULL, legend = "right")
}
\arguments{
\item{output_wvi}{WASABI output; must contain \code{particles} and \code{part.weights}.}

\item{Y}{Numeric matrix/data.frame with at least two columns.}

\item{title}{Optional title for the plot.}

\item{legend}{Where to put the legend ("right" or "bottom").}
}
\value{
A ggplot2 multi-facet plot object.
}
\description{
For each particle, plot its clustering in the PC1-PC2 space or first-two-feature space, in a facet.
}
\examples{
\dontrun{
ggscatter_grid2d(output_wvi, Y)
}
}
