% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ggscatter_pca}
\alias{ggscatter_pca}
\title{Grid of 2D cluster scatterplots from PCA coordinates}
\usage{
ggscatter_pca(output_wvi, Y, title = NULL)
}
\arguments{
\item{output_wvi}{WASABI output; must contain \code{particles} and \code{part.weights}.}

\item{Y}{Numeric matrix of dimension $n$ by $d$, with $d > 2$, containing the data points.}

\item{title}{Optional plot title.}
}
\value{
A ggplot2 multi-facet plot object.
}
\description{
Like \code{ggscatter_grid2d}, but uses principal component axes for the x/y axes.
}
\examples{
\dontrun{
ggscatter_pca(output_wvi, Y)
}
}
