% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{ggsummary}
\alias{ggsummary}
\title{Summary barplot for WASABI particles}
\usage{
ggsummary(output_wvi, title = NULL, legend = "right")
}
\arguments{
\item{output_wvi}{WASABI output; must contain \code{particles} and \code{part.weights}.}

\item{title}{Optional plot title.}

\item{legend}{Position of legend ("right" [default] or "bottom").}
}
\value{
A ggplot2 barplot object.
}
\description{
Plots both particle weights and corresponding number of clusters as a barplot for each sampled particle.
}
\examples{
\dontrun{
ggsummary(output_wvi)
}
}
