% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{omARI_compute_Rcpp}
\alias{omARI_compute_Rcpp}
\title{Compute "one minus adjusted Rand index" between two partitions}
\usage{
omARI_compute_Rcpp(c1, c2, K1, K2)
}
\arguments{
\item{c1}{An integer vector of cluster labels for $n$ items. Should be 0-indexed.}

\item{c2}{An integer vector of cluster labels for $n$ items. Should be 0-indexed.}

\item{K1}{An integer, specifying the number of unique clusters in \code{cl1}.}

\item{K2}{An integer, specifying the number of unique clusters in \code{cl2}.}
}
\value{
The VI distance between \code{cl1} and \code{cl2}.
}
\description{
Compute "one minus adjusted Rand index" between two partitions
}
