% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{relabel_partition}
\alias{relabel_partition}
\title{Relabel partitions}
\usage{
relabel_partition(c)
}
\arguments{
\item{c}{A vector of cluster labels of size $n$, where each label represents a cluster.}
}
\value{
A vector of the same size as \code{c}, where each cluster label is replaced with a number from 1 to K, where K is the number of unique clusters in \code{c}.
}
\description{
Change cluster labels so that they are ordered from 1 to K, 
where K is the number of clusters in the partition.
}
\examples{
\dontrun{
# assuming cls.draw is a matrix of size S x n, 
# where S is the number of samples and n is the sample size
# i.e. each row contains a partition of n data points
cls.draw = t(apply(cls.draw,1,relabel_partition)) 
}
}
