% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reorder_find_unique}
\alias{reorder_find_unique}
\title{Reorder cluster labels and find unique clusterings}
\usage{
reorder_find_unique(S, n, config)
}
\arguments{
\item{S}{Integer, number of samples.}

\item{n}{Integer, sample size.}

\item{config}{A matrix of size \code{S} x \code{n}, where each row represents a sample configuration.}
}
\value{
A list containing:
\begin{itemize}
\item \code{config_reorder} A matrix of size \code{S} x \code{n}, where each row represents a reordered sample configuration.
\item \code{config_count} A vector containing the count of each unique configuration (number of times each configuration in config_reorder appears in the original config.
\item \code{config_index} A vector containing the indices of unique configurations.
\end{itemize}
}
\description{
Reorder cluster labels in a collection of samples, so that cluster labels appear as 1,2,3,.. and to make
finding unique clusters more efficient.
}
