% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet.R
\name{vi.contribution}
\alias{vi.contribution}
\title{VI contribution between two partitions}
\usage{
vi.contribution(Z1, Z2)
}
\arguments{
\item{Z1}{Integer or factor vector: length-n cluster labels for partition 1.}

\item{Z2}{Integer or factor vector: length-n cluster labels for partition 2.}
}
\value{
Numeric vector of length n: VI contribution for each observation.
}
\description{
Calculates the individual contributions of observations to the VI distance between partitions Z1 and Z2.
}
