% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VI_Rcpp}
\alias{VI_Rcpp}
\title{Compute Variation of Information (VI) distance between two groups of partitions}
\usage{
VI_Rcpp(cls1, cls2, K1s, K2s, a = 1)
}
\arguments{
\item{cls1}{A matrix of $n$ columns, where each row contains the 0-indexed cluster labels for n items.}

\item{cls2}{A matrix of $n$ columns, where each row contains the 0-indexed cluster labels for n items.}

\item{K1s}{An integer vector, specifying the number of clusters in each row of \code{cls1}. Must have length equal to the number of rows in \code{cls1}.}

\item{K2s}{An integer vector, specifying the number of clusters in each row of \code{cls2}. Must have length equal to the number of rows in \code{cls2}.}

\item{a}{A parameter used in generalized VI, takes value between 0 and 2 and 1 by default.}
}
\value{
A matrix containing the VI distance between each row of \code{cls1} and \code{cls2}.
}
\description{
Compute Variation of Information (VI) distance between two groups of partitions
}
