% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WASABI.ext-package.R
\docType{package}
\name{WASABI.ext-package}
\alias{WASABI.ext}
\alias{WASABI.ext-package}
\title{WASABI.ext: What the Package Does (One Line, Title Case)}
\description{
WASABI provides a tool for summarizing the posterior distribution over clusterings in Bayesian mixture models. The package implements the Wasserstein Approximation for Bayesian clustering (WASABI) algorithm, which identifies multiple representative partitions (“particles”) that capture posterior uncertainty, by finding an approximation of the posterior that minimizes the Wasserstein distance based on the Variation of Information (VI) distance.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
