% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WD.R
\name{WD}
\alias{WD}
\title{Computes the Wassertein distance between the empirical posterior and the WASABI distribution}
\usage{
WD(
  cls.draw,
  part,
  part.weights,
  assign.vi,
  loss = c("VI", "Binder", "omARI"),
  a = 1
)
}
\arguments{
\item{cls.draw}{A matrix of the MCMC samples of partitions of $n$ data points.}

\item{part}{The WASABI particles. A matrix of dimension $L × n$ containing the particles obtained by WASABI in each of the rows.}

\item{part.weights}{A vector of length $L$ containing the weights associated to each of the particles.}

\item{assign.vi}{A vector containing the assignment of each sample \code{cls.draw} to a region of attraction (and the corresponding particle). It is obtained as output from WASABI.}

\item{loss}{Loss function used for WASABI approximation, can be chosen from "VI","Binder" and "omARI".}

\item{a}{Parameter for generalized VI and generalized Binder, takes value between 0 and 2, a = 1 by default (regular loss).}
}
\value{
The Wasserstein-VI distance between the empirical posterior distribution (supported on the MCMC samples) and the WASABI distribution.
}
\description{
Computes the Wassertein distance between the empirical posterior and the WASABI distribution
}
