% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet.R
\name{eb.contribution}
\alias{eb.contribution}
\title{Expected Binder Contribution}
\usage{
eb.contribution(Zmat, Zhat)
}
\arguments{
\item{Zmat}{Integer matrix: S rows (samples) by n columns (data points), each row a clustering.}

\item{Zhat}{Integer vector: length-n partition to compare.}
}
\value{
Numeric vector of length n: expected VI contribution for \code{Zhat} (each element).
}
\description{
Calculates the EB contribution of each observation to the expected Binder distance between a partition and samples.
}
