% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meet.R
\name{evi.contribution}
\alias{evi.contribution}
\title{Expected VI Contribution}
\usage{
evi.contribution(Zmat, Zhat, a = 1)
}
\arguments{
\item{Zmat}{Integer matrix: S rows (samples) by n columns (data points), each row a clustering.}

\item{Zhat}{Integer vector: length-n partition to compare.}

\item{a}{Parameter for generalized VI takes value between 0 and 2, a = 1 by default.}
}
\value{
Numeric vector of length n: expected VI contribution for \code{Zhat} (each element).
}
\description{
Calculates the EVI contribution of each observation to the expected variation of information (VI) distance between a partition and samples.
}
